+++ azuredevops/provider_test.go
@@ -140,7 +140,7 @@ func TestProvider_HasChildResources(t *testing.T) {
 		// NOTE: betterado_workitemtrackingprocess_page has been migrated to the framework provider.
 		// NOTE: betterado_workitemtrackingprocess_process has been migrated to the framework provider.
 		"betterado_workitemtrackingprocess_process_permissions",
-		"betterado_workitemtrackingprocess_rule",
+		// NOTE: betterado_workitemtrackingprocess_rule has been migrated to the framework provider.
 		"betterado_workitemtrackingprocess_system_control",
 		// NOTE: betterado_workitemtrackingprocess_state has been migrated to the framework provider.
 		// NOTE: betterado_workitemtrackingprocess_workitemtype has been migrated to the framework provider.